refactor: use new utils for forgotpassword and login views page object
